+++ apps/storybook/src/stories/Dialog.stories.tsx
@@ -19,7 +19,7 @@ export default {
     badges: [BADGE.STABLE],
     design: {
       type: 'figma',
-      url: '',
+      url: 'https://www.figma.com/file/nGRJSZ90AzGrFe23Shmkzf/Scala?type=design&node-id=638%3A167604&mode=design&t=QcKO9NKDxvGimjvK-1',
     },
   },
   args: {